requirements.txt
discord.py==1.7.3
openpyxl==3.0.10